Float animation indices fixed

+++ actors.py
@@ -17,7 +17,7 @@ factor=1000000
 def drawAnimation(screen,frames,pos,speed,t):
     if debug or debugFlags["anim"]:
         print(str(math.floor(pos[0]))+","+str(math.floor(pos[1]))+" anim pos for "+str(math.floor(speed*t%len(frames)))+" in "+str(frames))
-    screen.blit(frames[math.floor(speed*t%len(frames))],(math.floor(pos[0]),math.floor(pos[1])))
+    screen.blit(frames[int(math.floor(speed*t%len(frames)))],(math.floor(pos[0]),math.floor(pos[1])))
 
 def makeDeadAnim(anim):
     deadAnim=[img.copy() for img in anim]
